requirements.txt
pandas~=1.1.3
python-binance
argparse~=1.4.0
numpy~=1.19.2
requests~=2.25.0
python-dateutil~=2.8.1
setuptools~=50.3.1